Handle existing user with matching email but no matching oauth id
In tests, we found that the `login_as!` method was not behaving
as expected, but instead creating a new user with the same email
yet no OAuth identity. This caused failures in request specs that
relied on services created in a `before` block being visible to
a user also created in a `before` block.

This PR fixes the issue, which will also handle the real-world
edge case of an existing user unlinking their OAuth login, and then
trying to login with it again.