When changing a service slug, warn the user that this will invalidate any user data already in the user datastore.
The service slug is used as part of the key to retrieve data from the user datastore,
and also the first part of the domain to which the user's cookie will be scoped. So,
changing the service slug on a service with live user data in, will render any existing
data irretrievable.