Use activerecord session store
The change in auth0 to use variable length authenticity tokens meant
that the size of the token was blowing past what the browser was capable
of storing. As a result the users session would be effectively wiped
each time they authenticated.

Using the activerecord session store allows for the necessary space for
the additional length of user data and token.